add dataset and change langchain directory name to langchain_manager, incase of a name consuming in python

+++ app.py
@@ -3,13 +3,13 @@ from llama_index import download_loader, GPTVectorStoreIndex
 from llama_index import LLMPredictor, ServiceContext, LangchainEmbedding
 from langchain.llms import AzureOpenAI
 from langchain.embeddings.openai import OpenAIEmbeddings
+from llama_index import StorageContext, load_index_from_storage
 import os
 import pickle
 import streamlit as st
 
 import logging
 import sys
-
 
 logging.basicConfig(
     stream=sys.stdout, level=logging.DEBUG
@@ -49,43 +49,44 @@ def main() -> None:
     service_context = ServiceContext.from_defaults(
         llm_predictor=llm_predictor, embed_model=embedding
     )
-    download_loader("GithubRepositoryReader")
-    docs = None
-    if os.path.exists("docs/docs.pkl"):
-        with open("docs/docs.pkl", "rb") as f:
-            docs = pickle.load(f)
+    if os.path.exists("./dataset") and len(os.listdir("./dataset")) != 0:
+        storage_context = StorageContext.from_defaults(persist_dir="./dataset")
+    else:
+        download_loader("GithubRepositoryReader")
+        docs = None
+        if os.path.exists("docs/docs.pkl"):
+            with open("docs/docs.pkl", "rb") as f:
+                docs = pickle.load(f)
 
-    if docs is None:
-        github_client = GithubClient(os.getenv("GITHUB_TOKEN"))
-        loader = GithubRepositoryReader(
-            github_client,
-            owner="ctripcorp",
-            repo="x-pipe",
-            filter_directories=(
-                [".", "doc"],
-                GithubRepositoryReader.FilterType.INCLUDE,
-            ),
-            filter_file_extensions=([".md"], GithubRepositoryReader.FilterType.INCLUDE),
-            verbose=True,
-            concurrent_requests=10,
-        )
+        if docs is None:
+            github_client = GithubClient(os.getenv("GITHUB_TOKEN"))
+            loader = GithubRepositoryReader(
+                github_client,
+                owner="ctripcorp",
+                repo="x-pipe",
+                filter_directories=(
+                    [".", "doc"],
+                    GithubRepositoryReader.FilterType.INCLUDE,
+                ),
+                filter_file_extensions=([".md"], GithubRepositoryReader.FilterType.INCLUDE),
+                verbose=True,
+                concurrent_requests=10,
+            )
 
-        docs = loader.load_data(branch="master")
+            docs = loader.load_data(branch="master")
 
-        with open("docs/docs.pkl", "wb") as f:
-            pickle.dump(docs, f)
+            with open("docs/docs.pkl", "wb") as f:
+                pickle.dump(docs, f)
 
-    index = GPTVectorStoreIndex.from_documents(docs, service_context=service_context)
+        index = GPTVectorStoreIndex.from_documents(docs, service_context=service_context)
+        index.storage_context.persist(persist_dir="./dataset")
+        storage_context = index.storage_context
 
+    index = load_index_from_storage(storage_context=storage_context)
     query_engine = index.as_query_engine(service_context=service_context)
 
     query = st.text_input("X-Pipe Wiki 问题:")
     if query:
-        index = GPTVectorStoreIndex.from_documents(
-            docs, service_context=service_context
-        )
-
-        query_engine = index.as_query_engine(service_context=service_context)
         response = query_engine.query(query)
         st.write(response)
